flacng: Fix build

+++ src/flacng/metadata.cc
@@ -33,7 +33,7 @@
 
 static size_t read_cb(void *ptr, size_t size, size_t nmemb, FLAC__IOHandle handle)
 {
-    size_t read;
+    int64_t read;
 
     if (handle == NULL)
     {
